+++ CHANGELOG.md
@@ -10,7 +10,8 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 *   Chargeback transaction type.
     *   Add class `\UnzerSDK\Resources\TransactionTypes\Chargeback`.
     *   Add methods `\UnzerSDK\Unzer::fetchChargeback` and `\UnzerSDK\Unzer::fetchChargebackById` to fetch chargeback information.
-*   Add `\UnzerSDK\Resources\PaymentTypes\PayU` payment Type.
+*   Add `\UnzerSDK\Resources\PaymentTypes\PayU` payment type.
+*   Add example for PayU payment type.
 
 ## [3.2.0](https://github.com/unzerdev/php-sdk/compare/3.1.0..3.2.0)